Add parsers for cron segment types
This adds five different "parsers" for a cron segment, to handle the
various different values that could be supplied. They are:

- *Any*, which deals specifically with `*`, and lists all possible values
  for a segment
- *Exact*, which deals with exact integers as supplied (e.g. `1`)
- *List*, which deals with comma-separated lists of integers (e.g.
  `1,2,3`)
- *Range*, which deals with ranges (e.g. `1-3`)
- *Step*, which deals with "step" values (e.g. `1/3`)

In addition to the parsing functionality, there's some basic validation
based on a minimum and maximum value, and a helper method to get a given
parser based on an input string by picking the best parser for the job
based on a regular expression.